Enabled kwargs to pass padding_mode to the ImageLoader function and allowed appending transforms
Previous behavior for padding was simply use "constant" mode for filling padded pixels from outside the image, this would allow to use different modes such as "mirror".

Transforms were previously overwritten when added with `ZarrDataset.add_transform()`. With a new argument append, this changes to allow appending sequential transforms.
