+++ .gitignore
@@ -137,3 +137,4 @@ tests/test_zarrs/*
 tests/test_tiffs/*
 tests/*.png
 .vscode/settings.json
+docs/source/examples/tests.qmd
