+++ tests/test_zarrdataset.py
@@ -2,21 +2,13 @@ import pytest
 import os
 import shutil
 from pathlib import Path
-from unittest import mock
 from collections import OrderedDict
 from _collections_abc import dict_items
 
-import operator
-
 import zarrdataset as zds
-import importlib
-import tqdm
 import numpy as np
 
 from tests.utils import IMAGE_SPECS
-
-import torch
-from torch.utils.data import DataLoader, ChainDataset
 
 
 def input_target_dtype_fn(input, target):
@@ -124,38 +116,6 @@ def patch_sampler_specs(request):
     return patch_sampler, patch_size, allow_incomplete_patches
 
 
-@pytest.mark.parametrize("image_dataset_specs", [
-    IMAGE_SPECS[10],
-], indirect=["image_dataset_specs"])
-def test_compatibility_no_tqdm(image_dataset_specs):
-    with mock.patch.dict('sys.modules', {'tqdm': None}):
-        importlib.reload(zds._zarrdataset)
-
-        assert isinstance(object, type(zds._zarrdataset.tqdm).__bases__), \
-            (f"When `tqdm` is not installed, progress bars should be disabled,"
-             f" and a dummy interface derived from object should be used by "
-             f"ZarrDataset. Got an interface derived from "
-             f"{type(zds._zarrdataset.tqdm).__bases__} instead")
-
-        dataset_specs, _ = image_dataset_specs
-
-        dataset = zds._zarrdataset.ZarrDataset(
-            dataset_specs=dataset_specs,
-            progress_bar=True
-        )
-
-        try:
-            next(iter(dataset))
-
-        except Exception as e:
-            raise AssertionError(f"No exceptions where expected, got {e} "
-                                 f"instead.")
-
-    with mock.patch.dict('sys.modules', {'tqdm': tqdm}):
-        importlib.reload(zds._zarrdataset)
-        importlib.reload(zds)
-
-
 @pytest.mark.parametrize("dataset_spec_class", [
     zds.DatasetSpecs,
     zds.ImagesDatasetSpecs,
@@ -182,7 +142,7 @@ def test_DataSpecs(dataset_spec_class):
     assert isinstance(generic_specs["transforms"], OrderedDict),\
         (f"Expected transforms to be an ordered dicionary, got "
          f"{type(generic_specs['transforms'])} instead.")
-    
+
     assert isinstance(generic_specs.items(), dict_items),\
         (f"Expected DatasetSpecs to return `dict_items` when called `.items()`"
          f", got {type(generic_specs.items())} instead.")
@@ -266,8 +226,8 @@ def test_ZarrDataset(image_dataset_specs, shuffle, return_positions,
     for sample in ds:
         n_samples += 1
         if (any(["labels" in mode["modality"] for mode in dataset_specs])
-          or return_positions
-          or return_worker_id):
+           or return_positions
+           or return_worker_id):
             assert isinstance(sample, tuple), \
                 (f"When `return_positions`, `return_worker_id` or a labels "
                  f"dataset specification is passed to ZarrDataset, retrieved "
@@ -288,7 +248,7 @@ def test_ZarrDataset(image_dataset_specs, shuffle, return_positions,
 
         assert tuple(sample_array.shape) == tuple(specs["shape"]), \
             (f"Sample expected to have shape {tuple(specs['shape'])}, got "
-            f"{sample_array.shape} ({sample_array.dtype}) instead")
+             f"{sample_array.shape} ({sample_array.dtype}) instead")
 
         if "labels" in dataset_specs:
             expected_labels_shape = tuple(
@@ -300,337 +260,43 @@ def test_ZarrDataset(image_dataset_specs, shuffle, return_positions,
 
             assert tuple(labels_array.shape) == expected_labels_shape, \
                 (f"Labels expected to have shape {expected_labels_shape}, got "
-                f"{labels_array.shape} ({labels_array.dtype}) instead")
+                 f"{labels_array.shape} ({labels_array.dtype}) instead")
 
     assert n_samples > 0, ("Expected more than zero samples extracted from "
                            "this experiment.")
 
 
 @pytest.mark.parametrize(
-    "image_dataset_specs, patch_sampler_specs, shuffle, draw_same_chunk", [
-        (IMAGE_SPECS[10], (32, False), True, False),
-        (IMAGE_SPECS[10], (32, False), True, True),
-        (IMAGE_SPECS[10], (32, False), False, True),
+    "image_dataset_specs", [
+        IMAGE_SPECS[6],
     ],
-    indirect=["image_dataset_specs", "patch_sampler_specs"]
+    indirect=["image_dataset_specs"]
 )
-def test_patched_ZarrDataset(image_dataset_specs, patch_sampler_specs,
-                             shuffle,
-                             draw_same_chunk):
+def test_transforms_lists(image_dataset_specs):
     dataset_specs, specs = image_dataset_specs
-    patch_sampler, patch_size, allow_incomplete_patches = patch_sampler_specs
 
     ds = zds.ZarrDataset(
-        dataset_specs=dataset_specs,
-        shuffle=shuffle,
-        patch_sampler=patch_sampler,
-        draw_same_chunk=draw_same_chunk
+        dataset_specs=dataset_specs
     )
 
-    array_idx = 0
-    label_idx = 1
+    ds.add_transform(("images", ), zds.ToDtype(np.float64))
+    ds.add_transform(("images", ), zds.ToDtype(np.float64))
 
-    n_samples = 0
+    assert len(ds._transforms[(dataset_specs["modality"], )]) == 1, \
+        (f"Expected only one transform to be added to the dataset, got "
+         f"{len(ds._transforms)} instead.")
 
-    for sample in ds:
-        n_samples += 1
-        assert isinstance(sample, tuple), \
-            (f"When `return_positions`, `return_worker_id` or a labels dataset"
-             f" specification is passed to ZarrDataset, retrieved samples "
-             f"should be a tuple, got {type(sample)} instead.")
+    ds.add_transform(("images", ), zds.ToDtype(np.float64))
+    ds.add_transform(("images", ), zds.ToDtype(np.float32),
+                     append=True)
 
-        assert isinstance(sample[0], np.ndarray), \
-            (f"Sample should be a Numpy NDArray, got {type(sample[0])} "
-             f"instead.")
+    assert len(ds._transforms[(dataset_specs["modality"], )]) == 2, \
+        (f"Expected two transforms to be added to the dataset, got "
+         f"{len(ds._transforms)} instead.")
 
-        sample_array = sample[array_idx]
-        labels_array = sample[label_idx]
+    ds_iter = iter(ds)
+    sample = next(ds_iter)
 
-        expected_shape = tuple(
-            min(patch_size, specs["shape"][specs["source_axes"].index(ax)])
-            if ax in patch_sampler.spatial_axes else
-            specs["shape"][specs["source_axes"].index(ax)]
-            for ax in specs["source_axes"]
-        )
-
-        expected_labels_shape = tuple(
-            patch_size if ax in patch_sampler.spatial_axes else
-            specs["shape"]["YX".index(ax)]
-            for ax in "YX"
-        )
-
-        assert tuple(sample_array.shape) == expected_shape, \
-            (f"Sample expected to have shape {expected_shape}, got "
-             f"{sample_array.shape} instead")
-
-        assert tuple(labels_array.shape) == expected_labels_shape, \
-            (f"Labels expected to have shape {expected_labels_shape}, got "
-             f"{labels_array.shape} instead")
-
-    assert n_samples > 0, ("Expected more than zero samples extracted from "
-                           "this experiment.")
-
-    # Second iteration
-    n_samples = 0
-    for sample in ds:
-        n_samples += 1
-        assert isinstance(sample, tuple), \
-            (f"When `return_positions`, `return_worker_id` or a labels dataset"
-             f" specification is passed to ZarrDataset, retrieved samples "
-             f"should be a tuple, got {type(sample)} instead.")
-
-        assert isinstance(sample[0], np.ndarray), \
-            (f"Sample should be a Numpy NDArray, got {type(sample[0])} "
-             f"instead.")
-
-        sample_array = sample[array_idx]
-        labels_array = sample[label_idx]
-
-        expected_shape = tuple(
-            min(patch_size, specs["shape"][specs["source_axes"].index(ax)])
-            if ax in patch_sampler.spatial_axes else
-            specs["shape"][specs["source_axes"].index(ax)]
-            for ax in specs["source_axes"]
-        )
-
-        expected_labels_shape = tuple(
-            patch_size if ax in patch_sampler.spatial_axes else
-            specs["shape"]["YX".index(ax)]
-            for ax in "YX"
-        )
-
-        assert tuple(sample_array.shape) == expected_shape, \
-            (f"Sample expected to have shape {expected_shape}, got "
-             f"{sample_array.shape} instead")
-
-        assert tuple(labels_array.shape) == expected_labels_shape, \
-            (f"Labels expected to have shape {expected_labels_shape}, got "
-             f"{labels_array.shape} instead")
-
-    assert n_samples > 0, ("Expected more than zero samples extracted from "
-                           "this experiment.")
-
-
-@pytest.mark.parametrize(
-    "image_dataset_specs, patch_sampler_specs", [
-        (IMAGE_SPECS[10], (1024, True)),
-        (IMAGE_SPECS[10], (1024, False)),
-    ],
-    indirect=["image_dataset_specs", "patch_sampler_specs"]
-)
-def test_greater_patch_ZarrDataset(image_dataset_specs, patch_sampler_specs):
-    dataset_specs, specs = image_dataset_specs
-    patch_sampler, patch_size, allow_incomplete_patches = patch_sampler_specs
-
-    ds = zds.ZarrDataset(
-        dataset_specs=dataset_specs,
-        patch_sampler=patch_sampler
-    )
-
-    n_samples = 0
-    for _ in ds:
-        n_samples += 1
-
-    if allow_incomplete_patches:
-        assert n_samples > 0, ("Expected at elast one sample when patch"
-                               " size is greater than the image size, and"
-                               " `allow_incomplete_patches` is True.")
-    else:
-        assert n_samples == 0, ("Expected zero samples since requested patch"
-                                " size is greater than the image size, and"
-                                " `allow_incomplete_patches` is False.")
-
-
-@pytest.mark.parametrize(
-    "image_dataset_specs, patch_sampler_specs, shuffle, draw_same_chunk,"
-    "batch_size, num_workers", [
-        (IMAGE_SPECS[10], (32, False), True, False, 2, 2),
-        ([IMAGE_SPECS[10]] * 4, (32, False), True, True, 2, 3),
-        ([IMAGE_SPECS[10]] * 2, (32, False), True, True, 2, 3),
-    ],
-    indirect=["image_dataset_specs", "patch_sampler_specs"]
-)
-def test_multithread_ZarrDataset(image_dataset_specs, patch_sampler_specs,
-                                 shuffle,
-                                 draw_same_chunk,
-                                 batch_size,
-                                 num_workers):
-    dataset_specs, specs = image_dataset_specs
-
-    patch_sampler, patch_size, allow_incomplete_patches = patch_sampler_specs
-
-    ds = zds.ZarrDataset(
-        dataset_specs=dataset_specs,
-        shuffle=shuffle,
-        patch_sampler=patch_sampler,
-        draw_same_chunk=draw_same_chunk
-    )
-
-    dl = DataLoader(
-        ds,
-        batch_size=batch_size,
-        num_workers=num_workers,
-        worker_init_fn=zds.zarrdataset_worker_init_fn,
-        drop_last=True
-    )
-
-    array_idx = 0
-    label_idx = 1
-
-    n_samples = 0
-
-    for sample in dl:
-        n_samples += 1
-        assert isinstance(sample, list), \
-            (f"When a labels dataset specification is passed to ZarrDataset "
-             f"and used with PyTorch DataLoader, samples should be a list, got"
-             f" {type(sample)} instead.")
-
-        assert isinstance(sample[0], torch.Tensor), \
-            (f"Sample should be a PyTorch Tensor, got {type(sample[0])} "
-             f"instead.")
-
-        sample_array = sample[array_idx]
-        labels_array = sample[label_idx]
-
-        expected_shape = [batch_size] + [
-            min(patch_size, specs["shape"][specs["source_axes"].index(ax)])
-            if ax in patch_sampler.spatial_axes else
-            specs["shape"][specs["source_axes"].index(ax)]
-            for ax in specs["source_axes"]
-        ]
-
-        expected_labels_shape = [batch_size] + [
-            patch_size if ax in patch_sampler.spatial_axes else
-            specs["shape"]["YX".index(ax)]
-            for ax in "YX"
-        ]
-
-        assert all(map(operator.eq,
-                       sample_array.shape,
-                       expected_shape)), \
-            (f"Sample expected to have shape {expected_shape}, got "
-             f"{sample_array.shape} instead")
-
-        assert all(map(operator.eq,
-                       labels_array.shape,
-                       expected_labels_shape)), \
-            (f"Labels expected to have shape {expected_labels_shape}, got "
-             f"{labels_array.shape} instead")
-
-    assert n_samples > 0, ("Expected more than zero samples extracted from "
-                           "this experiment.")
-
-
-@pytest.mark.parametrize(
-    "image_dataset_specs, patch_sampler_specs, shuffle, draw_same_chunk,"
-    "batch_size, num_workers, repeat_dataset", [
-        (IMAGE_SPECS[10:12], (32, False), True, False, 2, 2, 1),
-        (IMAGE_SPECS[10:12], (32, False), True, False, 2, 2, 2),
-        (IMAGE_SPECS[10:12], (32, False), True, False, 2, 2, 3),
-    ],
-    indirect=["image_dataset_specs", "patch_sampler_specs"]
-)
-def test_multithread_chained_ZarrDataset(image_dataset_specs,
-                                         patch_sampler_specs,
-                                         shuffle,
-                                         draw_same_chunk,
-                                         batch_size,
-                                         num_workers,
-                                         repeat_dataset):
-    dataset_specs, specs = image_dataset_specs
-    patch_sampler, patch_size, allow_incomplete_patches = patch_sampler_specs
-
-    ds = [zds.ZarrDataset(dataset_specs=dataset_specs,
-                          shuffle=shuffle,
-                          patch_sampler=patch_sampler,
-                          draw_same_chunk=draw_same_chunk
-        )] * repeat_dataset
-
-    chained_ds = ChainDataset(ds)
-
-    dl = DataLoader(
-        chained_ds,
-        batch_size=batch_size,
-        num_workers=num_workers,
-        worker_init_fn=zds.chained_zarrdataset_worker_init_fn,
-        drop_last=True
-    )
-
-    array_idx = 0
-    label_idx = 1
-
-    n_samples = 0
-
-    for sample in dl:
-        n_samples += 1
-        assert isinstance(sample, list), \
-            (f"When a labels dataset specification is passed to ZarrDataset "
-             f"and used with PyTorch DataLoader, samples should be a list, got"
-             f" {type(sample)} instead.")
-
-        assert isinstance(sample[0], torch.Tensor), \
-            (f"Sample should be a PyTorch Tensor, got {type(sample[0])} "
-             f"instead.")
-
-        sample_array = sample[array_idx]
-        labels_array = sample[label_idx]
-
-        expected_shape = [batch_size] + [
-            min(patch_size, specs["shape"][specs["source_axes"].index(ax)])
-            if ax in patch_sampler.spatial_axes else
-            specs["shape"][specs["source_axes"].index(ax)]
-            for ax in specs["source_axes"]
-        ]
-
-        expected_labels_shape = [batch_size] + [
-            patch_size if ax in patch_sampler.spatial_axes else
-            specs["shape"]["YX".index(ax)]
-            for ax in "YX"
-        ]
-
-        assert all(map(operator.eq,
-                       list(sample_array.shape),
-                       expected_shape)), \
-            (f"Sample expected to have shape {expected_shape}, got "
-             f"{sample_array.shape} instead")
-
-        assert all(map(operator.eq,
-                       labels_array.shape,
-                       expected_labels_shape)), \
-            (f"Labels expected to have shape {expected_labels_shape}, got "
-             f"{labels_array.shape} instead")
-
-    assert n_samples > 0, ("Expected more than zero samples extracted from "
-                           "this experiment.")
-
-
-def test_compatibility_no_pytroch():
-    with mock.patch.dict('sys.modules', {'torch': None}):
-        importlib.reload(zds._zarrdataset)
-
-        dataset = zds._zarrdataset.ZarrDataset()
-
-        assert isinstance(object, type(dataset).__bases__), \
-            (f"When pytorch is not installed, ZarrDataset should be inherited"
-             f" from object, not {type(dataset).__bases__}")
-
-        try:
-            zds._zarrdataset.zarrdataset_worker_init_fn(None)
-
-        except Exception as e:
-            raise AssertionError(f"No exceptions where expected when using "
-                                 f"`zarrdataset_worker_init_fn` without "
-                                 f"pytorch installed, got {e} instead.")
-
-        try:
-            zds._zarrdataset.chained_zarrdataset_worker_init_fn(None)
-
-        except Exception as e:
-            raise AssertionError(f"No exceptions where expected when using "
-                                 f"`chained_zarrdataset_worker_init_fn` "
-                                 f"without pytorch installed, got {e} "
-                                 f"instead.")
-    with mock.patch.dict('sys.modules', {'torch': torch}):
-        importlib.reload(zds._zarrdataset)
-        importlib.reload(zds)
+    assert sample[0].dtype == np.float32, \
+        (f"Expected sample to have data type numpy.float32 after applying two "
+         f"consecutive data type transforms, got {sample[0].dtype} instead.")
